Numerous small fixes, clean up output a bit, implement firewall funciton
The firewall function has been implemented in this version. It will read from "./firewall.txt" and store the rules as a string in an arraylist like so: "(3,*) :local" (it was easier for me to work with) for the switches and nodes to check if they are involved in the rule. It seems to work well based on the output files (i wouldn't trust all the txt that gets printed out but it will show that the firewall function is working, you won't see it often because the cases don't always occur, try it with different txt generations)